06_PipeLine/requirements.txt
opencv-python==4.4.0.46
tensorflow==2.4.0
keras==2.4.0
Pillow==8.0.1
config
sklearn
numpy
pandas
matplotlib
pytest-shutil
scikit-image
